fix(FEV-1011): merge VOD provider
