+++ src/types.ts
@@ -6,3 +6,16 @@ export const CuepointType: Record<string, string> = {
   SLIDE: 'slide',
   HOTSPOT: 'hotspot'
 };
+
+export enum KalturaThumbCuePointSubType {
+  CHAPTER = 2,
+  SLIDE = 1
+}
+export enum KalturaCuePointType {
+  // All: 'All',
+  // AnswersOnAir: 'AnswersOnAir',
+  // Chapters: 'Chapters',
+  SLIDE = 'slide'
+  // Hotspots: 'Hotspots',
+  // Captions: 'Captions'
+}
